Minor file/namespace reorganization
