+++ PalCalc.UI/ViewModel/Inspector/Search/Grid/DefaultContainerGridViewModel.cs
@@ -13,7 +13,7 @@ using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 
-namespace PalCalc.UI.ViewModel.Inspector.Search
+namespace PalCalc.UI.ViewModel.Inspector.Search.Grid
 {
     public partial class DefaultContainerGridViewModel(List<PalInstance> contents) : ObservableObject, IContainerGridViewModel
     {
